Keep track of distance travelled
To solve part two I need to know how far it takes to reach each
intersection, to determine the shortest path. This commit implements an
additional field, total_distance to track how far each point it from the
beginning. Since the distance is not relevant to finding the
intersection the Hash and PartialEq traits had to be manually
implemented to exclude the total_distance property from being included
in the comparison logic.