Stop including central point in list of points
The vector was originally initialised with the central point to get
around the issue of being empty when processing the first movement,
which also required it to be removed later, before determining the
manhattan distance. This commit updates the code to unwrap or provide
the central point as a default, so it no longer gets included and has to
be explicitly removed.