Cache intermediate states
The current script takes about ~18 seconds to iterate. With storing the
value of the number -> string conversion in it's own value the average
execution has reduced to ~12 seconds. This is still pretty long for what
the file is doing, and I know for a fact the code is real messy, so
there are much better performance gains to be had.