Change O(n) string access to O(1)
.chars().nth(i) is O(n) because of UTF-8. Since we know that the string
is ASCII, we can use bytewise access instead, which allows O(1) access.
.bytes().nth(i) can theoretically be optimised from O(n) to O(1), but
I’m not sure if that happens in general, so I’m going the “safe” route
with .as_bytes().get(i).cloned() and the likes. (.get() produces
Option<&u8>, .cloned() turns that into Option<u8>.) A fairly messy way
of writing it still, which could be improved, but it’s enough.